finish detail->util rename, better keyword search
