+++ include/librepr/parsing/token.h
@@ -7,32 +7,34 @@
 
 #include <librepr/util/concepts.h>
 #include "token_kind.h"
-#include "lex_error.h"
+#include "token/lex_error.h"
 
 namespace librepr::parsing {
 
+enum class TokenCategory : unsigned char { error, generic, character, string, numeric, name };
+
 template <auto V>
-concept is_token_category = std::same_as<std::remove_const_t<decltype(V)>, TokenCategory::Category>;
+concept is_token_category = std::same_as<std::remove_const_t<decltype(V)>, TokenCategory>;
 
 class Token {
 private:
 public:
-  TokenCategory::Category category{TokenCategory::error};
   union Type {
     LexError::Error lex_error;
     TokenKind::Kind kind;
     CharacterFlags::Flag char_flags;
     StringFlags::Flag string_flags;
     Numeral numeric_flags;
-    Name name;
+    detail::Name name_flags;
 
-    static_assert(EnableUnion<&Type::lex_error,
-                              &Type::kind,
-                              &Type::char_flags,
-                              &Type::string_flags,
-                              &Type::numeric_flags,
-                              &Type::name>);
+    static_assert(util::EnableUnion<util::UnionMember{&Type::lex_error, TokenCategory::error},
+                                    util::UnionMember{&Type::kind, TokenCategory::generic},
+                                    util::UnionMember{&Type::char_flags, TokenCategory::character},
+                                    util::UnionMember{&Type::string_flags, TokenCategory::string},
+                                    util::UnionMember{&Type::numeric_flags, TokenCategory::numeric},
+                                    util::UnionMember{&Type::name_flags, TokenCategory::name}>);
   } type{.lex_error = LexError::Unknown};
+  TokenCategory category{TokenCategory::error};
 
 public:
   uint16_t start;
@@ -40,75 +42,44 @@ public:
   constexpr Token() : start(0), end(0) {}
   constexpr Token(uint16_t start_, uint16_t end_) : start(start_), end(end_) {}
 
-  /**
-   * @brief Construct from class type union members
-   *
-   * @tparam T A class type that's a member of the Type union
-   * @param start_
-   * @param end_
-   * @param type_
-   */
-  template <librepr::util::is_class T>
-    requires is_token_category<T::tag>
-  constexpr Token(uint16_t start_, uint16_t end_, T type_) : start(start_),
-                                                             end(end_),
-                                                             category(T::tag) {
-    std::construct_at(&(type.*get_union_accessor<Type, T::tag>), type_);
+  template <typename T>
+    requires util::is_tagged_union<Type, T>
+  constexpr explicit Token(uint16_t start_, uint16_t end_, T value)
+      : start(start_)
+      , end(end_)
+      , category(util::get_union_tag<Type, T>) {
+    std::construct_at(&(type.*util::get_union_accessor<Type, util::get_union_tag<Type, T>>), value);
   }
 
-  /**
-   * @brief Construct from enum type union members
-   *
-   * @tparam T An enum type that's a member of the Type union
-   * @param start_
-   * @param end_
-   * @param type_
-   */
-  template <librepr::util::is_enum T>
-    requires is_token_category<get_tag<T>>
-  constexpr Token(uint16_t start_, uint16_t end_, T type_) : start(start_),
-                                                             end(end_),
-                                                             category(get_tag<T>) {
-    std::construct_at(&(type.*get_union_accessor<Type, get_tag<T>>), type_);
-  }
-
-  template <librepr::util::is_class T>
-    requires is_token_category<T::tag>
-  constexpr Token& operator=(T type_) {
-    category = T::tag;
-    std::construct_at(&(type.*get_union_accessor<Type, T::tag>), type_);
-    return *this;
-  }
-
-  template <librepr::util::is_enum T>
-    requires is_token_category<get_tag<T>>
-  constexpr Token& operator=(T type_) {
-    category = get_tag<T>;
-    std::construct_at(&(type.*get_union_accessor<Type, get_tag<T>>), type_);
+  template <typename T>
+    requires util::is_tagged_union<Type, T>
+  constexpr Token& operator=(T value) {
+    category = util::get_union_tag<Type, T>;
+    std::construct_at(&(type.*util::get_union_accessor<Type, util::get_union_tag<Type, T>>), value);
     return *this;
   }
 
   template <typename T>
   [[nodiscard]] constexpr T& get() {
-    constexpr TokenCategory::Category tag = get_tag<T>;
+    constexpr TokenCategory tag = util::get_union_tag<Type, T>;
 
     if (tag == category) {
-      return type.*get_union_accessor<Type, tag>;
+      return type.*util::get_union_accessor<Type, tag>;
     }
     throw std::runtime_error("Member not held");
   }
 
   template <typename T>
   [[nodiscard]] constexpr T const& get() const {
-    constexpr TokenCategory::Category tag = get_tag<T>;
+    constexpr TokenCategory tag = util::get_union_tag<Type, T>;
 
     if (tag == category) {
-      return type.*get_union_accessor<Type, tag>;
+      return type.*util::get_union_accessor<Type, tag>;
     }
     throw std::runtime_error("Member not held");
   }
 
-  [[nodiscard]] constexpr bool is(TokenCategory::Category category_) const { return category == category_; }
+  [[nodiscard]] constexpr bool is(TokenCategory category_) const { return category == category_; }
 
   template <librepr::util::is_class T>
     requires is_token_category<T::tag>
@@ -124,19 +95,19 @@ public:
   template <typename T>
     requires std::is_enum_v<T>
   [[nodiscard]] constexpr bool is(T flags) const {
-    if (category != get_tag<T>) {
+    if (category != util::get_union_tag<Type, T>) {
       return false;
     }
 
-    auto const& current = get<get_union_type<Type, get_tag<T>>>();
+    auto const& current = get<util::union_member_type<Type, T>>();
     static_assert(std::same_as<std::remove_cvref_t<decltype(current)>, T>);
 
     return current == flags;
   }
 
-  [[nodiscard]] constexpr bool in(TokenCategory::Category category_) const { return category_ == category; }
+  [[nodiscard]] constexpr bool in(TokenCategory category_) const { return category_ == category; }
 
-  template <std::same_as<TokenCategory::Category>... Ts>
+  template <std::same_as<TokenCategory>... Ts>
   [[nodiscard]] constexpr bool in(Ts... alternatives) const {
     return ((category == alternatives) || ...);
   }
@@ -153,13 +124,13 @@ public:
   }
 
   template <librepr::util::is_enum T, std::same_as<T>... Ts>
-    requires(!std::same_as<T, TokenCategory::Category>)
+    requires(!std::same_as<T, TokenCategory>)
   [[nodiscard]] constexpr bool in(T flag, Ts... flags) const {
-    if (category != get_tag<T>) {
+    if (category != util::get_union_tag<Type, T>) {
       return false;
     }
 
-    auto const& current = get<get_union_type<Type, get_tag<T>>>();
+    auto const& current = get<util::union_member_type<Type, T>>();
     static_assert(std::same_as<std::remove_cvref_t<decltype(current)>, T>);
     return (current & static_cast<T>((flag | ... | flags))) != 0;
   }
